# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend vor Rupay",
        "title": "Rupay Backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/api",
    "paths": {
...
        "/event": {
            "post": {
                "description": "Create a new event with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Create a new event",
                "parameters": [
                    {
                        "description": "CreateEvent payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateEvent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event data",
                        "schema": {
                            "$ref": "#/definitions/db.Event"
                        }
                    },
                    "400": {
                        "description": "Invalid Payload",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/event/{id}": {
            "get": {
                "description": "Retrieve an event by the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Retrieve an event by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event data",
                        "schema": {
                            "$ref": "#/definitions/db.Event"
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an event by the provided ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Delete an event by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Event deleted successfully"
                    },
                    "404": {
                        "description": "Event not found",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update an event by the provided id and details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Update an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event data",
                        "schema": {
                            "$ref": "#/definitions/db.Event"
                        }
                    },
                    "400": {
                        "description": "Invalid Payload",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events": {
            "get": {
                "description": "Retrieve a list of all events",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Retrieve all events",
                "responses": {
                    "200": {
                        "description": "List of events",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Event"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.Event": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "from_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "to_date": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "e.ErrorDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "Field causing the error",
                    "type": "string"
                },
                "issue": {
                    "description": "Description of the issue with the field",
                    "type": "string"
                }
            }
        },
        "e.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "string"
                },
                "details": {
                    "description": "List of field-specific errors",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/e.ErrorDetail"
                    }
                },
                "error": {
                    "description": "Stack Trace",
                    "type": "string"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string"
                }
            }
        },
...
        "schemas.CreateEvent": {
            "type": "object",
            "required": [
                "from_date",
                "name",
                "to_date"
            ],
            "properties": {
                "desc": {
                    "type": "string"
                },
                "from_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "to_date": {
                    "type": "string"
                }
            }
        },
...
    }
}